Fixed several bugs and removed repetitive code
There were a few issues with the update logic.
Furthermore, the code now behaves identically for the positive
and negative arrays, which greatly improves readability, removes
repetitive code. The only drawback is that the negative 0 index
is now not used, which slighty increases the memory consumption,
but results in code that is much easier to debug and understand.