Make the map structure observable
The map class now extends Observable, which allows anyone who
is interested in changes to the map, to implement the Observer
interface. This is especially handy for the GUI, as it also provides
the x,y and the value of the updated cell, without having to query
the map. The observer pattern can also come in handy if we decide to
create an optimised map.