Convert Gui class to a JPanel and rename
Instead of offering a complete JFrame, the class now offers a custom
JPanel, with example JFrame in the main method. The provided example
places the custom component into another JPanel, in order to center it
on to the screen. A scroll pane is also added, in case the map grows
larger than the containing JFrame.