Encode reading distance in each cell.
Each cell of the map now, not only contains the TYPE of each cell,
but also the DISTANCE from which the TYPE was determined/sensed.

This additional information is used during sensing, so each cell can be
updated, but only if observed from a closer distance than the last time.

Other classes modified to take care of modified signatures.