Only draw unoptimised path, draw end point.
As the graphical tools have access to both paths, and they are
geometrically identical, draw only the unoptimised path, so Bresenham
doesn't have to be used to fill in the gaps.

The end point of the path changes color, if it changes state from
unexplored to explored, while still following the path. This helps
with optimising the do360SonarScan, and seeing when the cell is explored.