Optimise do360Sonar scanning during exploration
Now, the turning threshold is smaller, and the robot turns more than 11.5
degrees, to make sure it has scanned 360 degrees.

If the do360Sonar scan fails to explore the cell of interest, it sets a
flag, so it is NOT called again until the next point on the path.

Status messages are printed to both consoles, to show when the method is
called and whether it managed to explore the cell.