Change sensing to only update cells if observed from closer distance.
This way, the closer the robot gets to a cell, the more reliable the
information about the cell is. As the robot moves away from the cell,
the information is not overwritten with less accurate information.

If the cell is currently BUFFER, but we want to change it to EMPTY,
we preserve the BUFFER state, but update the distance, if closer.